tweaked daemon make file, fixed node message handler
Fixed compile error in Windows for lack of a build.h file. Made PoS staking more in line with original Yacoin and less nvc or ppc-like. adjusted some refresh constants for a smoother start up display.  Slightly changed the daemon Windows build file for better behavior wrt the leveldb files.  The main change is to the node protocol handler, which was aborting sometimes in Windows,.  Added a total up time item, visible in the rpc 'getinfo'.